# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HackForge Hackathon Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface VerifyCodeBody {
  code: string;
}

export interface AdminLoginBody {
  email: string;
  password: string;
}

export interface AuthSession {
  token: string;
  participantCode: string;
  hasVoted: boolean;
}

export interface AdminAuthSession {
  token: string;
  email: string;
  isAdmin: boolean;
}

export interface MeResponse {
  isAdmin: boolean;
  participantCode?: string;
  hasVoted?: boolean;
}

export interface ParticipationCode {
  id: number;
  code: string;
  isUsed: boolean;
  usedAt?: string | null;
  createdAt: string;
}

export interface GenerateCodesBody {
  /**
   * @minimum 1
   * @maximum 500
   */
  count: number;
}

export interface CodeStats {
  total: number;
  used: number;
  unused: number;
}

export interface Team {
  id: number;
  name: string;
  projectTitle: string;
  description?: string | null;
  githubUrl?: string | null;
  createdAt: string;
}

export interface CreateTeamBody {
  name: string;
  projectTitle: string;
  description?: string | null;
  githubUrl?: string | null;
}

export interface LeaderboardEntry {
  teamId: number;
  teamName: string;
  projectTitle: string;
  voteCount: number;
  percentage: number;
}

export interface Poll {
  id: number;
  question: string;
  isActive: boolean;
  isFrozen: boolean;
  createdAt: string;
  totalVotes: number;
}

export interface PollResult {
  teamId: number;
  teamName: string;
  projectTitle: string;
  voteCount: number;
  percentage: number;
}

export interface PollWithResults {
  id: number;
  question: string;
  isActive: boolean;
  isFrozen: boolean;
  totalVotes: number;
  results: PollResult[];
}

export interface CreatePollBody {
  question: string;
}

export interface CastVoteBody {
  teamId: number;
}

export interface Vote {
  id: number;
  teamId: number;
  teamName: string;
  pollId: number;
  createdAt: string;
}

export interface MyVoteStatus {
  hasVoted: boolean;
  votedForTeamId?: number | null;
  votedForTeamName?: string | null;
}

export interface TeamResult {
  rank: number;
  teamId: number;
  teamName: string;
  projectTitle: string;
  voteCount: number;
  percentage: number;
}

export interface PublicResults {
  isPublished: boolean;
  publishedAt?: string | null;
  winner?: TeamResult | null;
  rankedTeams: TeamResult[];
}

export type EventStatusPhase =
  (typeof EventStatusPhase)[keyof typeof EventStatusPhase];

export const EventStatusPhase = {
  registration: "registration",
  submission: "submission",
  shortlisting: "shortlisting",
  elimination: "elimination",
  finale: "finale",
  results: "results",
  closed: "closed",
} as const;

export interface EventStatus {
  phase: EventStatusPhase;
  streamUrl?: string | null;
  streamActive: boolean;
  resultsPublished: boolean;
  eventName: string;
  tagline: string;
  updatedAt: string;
}

export type UpdateEventStatusBodyPhase =
  (typeof UpdateEventStatusBodyPhase)[keyof typeof UpdateEventStatusBodyPhase];

export const UpdateEventStatusBodyPhase = {
  registration: "registration",
  submission: "submission",
  shortlisting: "shortlisting",
  elimination: "elimination",
  finale: "finale",
  results: "results",
  closed: "closed",
} as const;

export interface UpdateEventStatusBody {
  phase?: UpdateEventStatusBodyPhase;
  streamUrl?: string | null;
  streamActive?: boolean;
  resultsPublished?: boolean;
  eventName?: string;
  tagline?: string;
}

export interface AdminDashboard {
  totalCodes: number;
  usedCodes: number;
  unusedCodes: number;
  totalTeams: number;
  totalVotes: number;
  activeParticipants: number;
  activePollQuestion?: string | null;
  currentPhase: string;
  streamActive: boolean;
}

export interface AdminLog {
  id: number;
  action: string;
  details?: string | null;
  createdAt: string;
}