# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HackForge Hackathon Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/verify-code:
    post:
      operationId: verifyParticipationCode
      tags: [auth]
      summary: Verify a participation code and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCodeBody"
      responses:
        "200":
          description: Code verified, session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "400":
          description: Invalid or already-used code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Admin session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session info
      responses:
        "200":
          description: Current user session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /codes:
    get:
      operationId: listCodes
      tags: [codes]
      summary: List all participation codes (admin)
      responses:
        "200":
          description: List of codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParticipationCode"
    post:
      operationId: generateCodes
      tags: [codes]
      summary: Generate participation codes (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCodesBody"
      responses:
        "201":
          description: Codes generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParticipationCode"
  /codes/{code}/reset:
    post:
      operationId: resetCode
      tags: [codes]
      summary: Reset a participation code (admin)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Code reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParticipationCode"
  /codes/{code}:
    delete:
      operationId: deleteCode
      tags: [codes]
      summary: Delete a participation code (admin)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Code deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /polls:
    get:
      operationId: listPolls
      tags: [polls]
      summary: List all polls (admin)
      responses:
        "200":
          description: List of polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poll"
    post:
      operationId: createPoll
      tags: [polls]
      summary: Create a new poll (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePollBody"
      responses:
        "201":
          description: Poll created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
  /polls/active:
    get:
      operationId: getActivePoll
      tags: [polls]
      summary: Get the currently active poll
      responses:
        "200":
          description: Active poll with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollWithResults"
        "404":
          description: No active poll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /polls/{id}/activate:
    post:
      operationId: activatePoll
      tags: [polls]
      summary: Activate a poll (admin — deactivates any other active poll)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Poll activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
  /polls/{id}/deactivate:
    post:
      operationId: deactivatePoll
      tags: [polls]
      summary: Deactivate a poll (admin — freeze results)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Poll deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Poll"
  /polls/{id}/results:
    get:
      operationId: getPollResults
      tags: [polls]
      summary: Get poll results with vote breakdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Poll results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollWithResults"

  # ── VOTES ─────────────────────────────────────────────────────────────────
...
  /event/status:
    get:
      operationId: getEventStatus
      tags: [event]
      summary: Get current event status and stream URL
      responses:
        "200":
          description: Event status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventStatus"
    put:
      operationId: updateEventStatus
      tags: [event]
      summary: Update event status and stream URL (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventStatusBody"
      responses:
        "200":
          description: Event status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventStatus"

  # ── ADMIN DASHBOARD ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    VerifyCodeBody:
      type: object
      properties:
        code:
          type: string
      required: [code]
    AdminLoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthSession:
      type: object
      properties:
        token:
          type: string
        participantCode:
          type: string
        hasVoted:
          type: boolean
      required: [token, participantCode, hasVoted]
    AdminAuthSession:
      type: object
      properties:
        token:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
      required: [token, email, isAdmin]
    MeResponse:
      type: object
      properties:
        isAdmin:
          type: boolean
        participantCode:
          type: string
        hasVoted:
          type: boolean
      required: [isAdmin]
    ParticipationCode:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        isUsed:
          type: boolean
        usedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, code, isUsed, createdAt]
    GenerateCodesBody:
      type: object
      properties:
        count:
          type: integer
          minimum: 1
          maximum: 500
      required: [count]
...
    Poll:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        isActive:
          type: boolean
        isFrozen:
          type: boolean
        createdAt:
          type: string
          format: date-time
        totalVotes:
          type: integer
      required: [id, question, isActive, isFrozen, createdAt, totalVotes]
    PollWithResults:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        isActive:
          type: boolean
        isFrozen:
          type: boolean
        totalVotes:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/PollResult"
      required: [id, question, isActive, isFrozen, totalVotes, results]
    PollResult:
      type: object
      properties:
        teamId:
          type: integer
        teamName:
          type: string
        projectTitle:
          type: string
        voteCount:
          type: integer
        percentage:
          type: number
      required: [teamId, teamName, projectTitle, voteCount, percentage]
    CreatePollBody:
      type: object
      properties:
        question:
          type: string
      required: [question]
...
    EventStatus:
      type: object
      properties:
        phase:
          type: string
          enum: [registration, submission, shortlisting, elimination, finale, results, closed]
        streamUrl:
          type: string
          nullable: true
        streamActive:
          type: boolean
        resultsPublished:
          type: boolean
        eventName:
          type: string
        tagline:
          type: string
        updatedAt:
          type: string
          format: date-time
      required: [phase, streamActive, resultsPublished, eventName, tagline, updatedAt]
    UpdateEventStatusBody:
      type: object
      properties:
        phase:
          type: string
          enum: [registration, submission, shortlisting, elimination, finale, results, closed]
        streamUrl:
          type: string
          nullable: true
        streamActive:
          type: boolean
        resultsPublished:
          type: boolean
        eventName:
          type: string
        tagline:
          type: string
...